add base entity
