+++ homeassistant/components/sleep_as_android/event.py
@@ -10,14 +10,13 @@ from homeassistant.components.event import (
     EventEntity,
     EventEntityDescription,
 )
-from homeassistant.const import CONF_WEBHOOK_ID
 from homeassistant.core import HomeAssistant, callback
-from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import SleepAsAndroidConfigEntry
-from .const import ATTR_EVENT, DOMAIN, SLEEP_AS_ANDROID_EVENT
+from .const import ATTR_EVENT, SLEEP_AS_ANDROID_EVENT
+from .entity import SleepAsAndroidEntity
 
 PARALLEL_UPDATES = 0
 
@@ -133,28 +132,11 @@ async def async_setup_entry(
     )
 
 
-class SleepAsAndroidEventEntity(EventEntity):
+class SleepAsAndroidEventEntity(SleepAsAndroidEntity, EventEntity):
     """An event entity."""
 
     _attr_has_entity_name = True
     entity_description: SleepAsAndroidEventEntityDescription
-
-    def __init__(
-        self,
-        config_entry: SleepAsAndroidConfigEntry,
-        entity_description: SleepAsAndroidEventEntityDescription,
-    ) -> None:
-        """Initialize the entity."""
-
-        self._attr_unique_id = f"{config_entry.entry_id}_{entity_description.key}"
-        self.entity_description = entity_description
-        self.webhook_id = config_entry.data[CONF_WEBHOOK_ID]
-        self._attr_device_info = DeviceInfo(
-            connections={(DOMAIN, config_entry.entry_id)},
-            manufacturer="Urbandroid",
-            model="Sleep as Android",
-            name=config_entry.title,
-        )
 
     @callback
     def _async_handle_event(self, webhook_id: str, data: dict[str, str]) -> None:
